+++ mayan/apps/documents/tests/test_trashed_document_views.py
@@ -17,7 +17,6 @@ class DocumentTrashViewTestCase(
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
 
-
         response = self._request_test_document_trash_get_view()
         self.assertEqual(response.status_code, 404)